Update bot movement path and fix mapId serialization type
Changed bot movement to use Z axis instead of Y in bot_profiles.h. Updated mapId serialization in player.cpp from uint64 to uint32 to match protocol requirements.

+++ src/realm_server/player.cpp
@@ -2371,7 +2371,7 @@ namespace mmo
 									   {
 			outPacket.Start(game::realm_client_packet::LoginVerifyWorld);
 			outPacket
-				<< io::write<uint64>(m_characterData->mapId)	
+				<< io::write<uint32>(m_characterData->mapId)
 				<< io::write<float>(m_characterData->position.x)
 				<< io::write<float>(m_characterData->position.y)
 				<< io::write<float>(m_characterData->position.z)
